Looks like we replaced DOM2DTM with DOM2DTMExt when we fixed
https://github.com/sparklemotion/nokogiri/pull/1320 but forgot to replace it in
the DOM2DTM manager

fixes #1749