@jeremyhaile 2018-09-05
@jvshahid thank you so much for this fix. I can confirm it fixes the memory issues with nested XPATH searching. However, there is still a pretty large and possibly exponential performance penalty for nested XPATH searches.

Here is the test script I am using: 

```
require 'nokogiri'

items = %Q(
<div class="test">
  <a href="something">derp</a>
</div>
) * ARGV[0].to_i

html = %Q(
<html>
  <body>
    #{items}
  </body>
</html>
)
doc = Nokogiri::HTML html
puts "pid is #{$$}"

start = Time.now
links = doc.css('html body .test a')
puts "Without nesting got #{links.length} links in #{Time.now - start}s" # This will return immediately

start = Time.now
links = doc.css('html body .test').css('a')
puts "With nesting got #{links.length} links in #{Time.now - start}s" # This will take a LONG time to finish
```

Here is the output with 5000 divs:
```
pid is 49017
Without nesting got 5000 links in 0.225391s
With nesting got 5000 links in 2.666568s
```

And here is the output with 50,000 divs:
```
pid is 49113
Without nesting got 50000 links in 0.7900900000000001s
With nesting got 50000 links in 296.062414s
```

So in both cases the nested version is several orders of magnitude slower than the unnested version. And it seems to get exponentially worse for larger numbers of matched elements.

This should definitely be merged in to fix the memory issue, as crashing the JVM is much worse than being slow. However there is still something wrong with nokogiri java's handling of nested XPATH queries.

Any idea why the performance might be so bad? Thanks again!