@jvshahid 2018-09-06
I think this is a result of the `NodeSet` implementation in JRuby which uses RubyArray under the hood. This causes the exponential behavior you are seeing as it tries to calculate the union in `.css('a')` for each node in the result of `.css('html body .test')`. As a workaround you can try using some variation of the following instead:

```
links = doc.css('html body .test').map do |node|
  node.css('a')
end
puts "finished"
links = Nokogiri::XML::NodeSet.new doc, links.flatten
```

I will probably open a new issue to track the exponential behavior of `NodeSet` and merge this PR to close #1749