@jeremyhaile 2018-09-06
@jvshahid thanks for the explanation and workaround. Unfortunately for us the badly performant code is in a third-party library that uses nokogiri, although we could fork it and modify. 

Please let me know when you open the new issue and I'll subscribe there too. Thanks again!