@jvshahid 2018-09-11
Just a quick update. I looked at libxml source code and it looks like internally it is using an array similar to the current implementation in JRuby. I am still not sure why the JRuby implementation is that slow.