@jvshahid 2018-09-15
created https://github.com/sparklemotion/nokogiri/issues/1795 to track the performance of XMLNodeSet. This PR is ready to merge AFAIK.